+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
   * New maintainer: Jiří Janoušek. Issue: tiliado/nuvola-app-soundcloud#11
   * Switched to use Chromium backed for playback without Flash plugin.
   * Metadata parsing was updated to reflect changes of the website.
+  * The track progress bar and volume bar were integrated.
 
 1.3 - February 12th, 2017
 -------------------------